+++ app/src/main/java/com/example/final_project/util/Order.kt
@@ -1,9 +1,10 @@
 package com.example.final_project.util
 
 data class Order (
+
     val uid: String,
     val foodname : String,
     val time : String,
     val amount : Int,
-    val total_price : Double
+    val total_price : Float
 )
